web-animations/animation-player/animation-player-reverse.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Animation player reverse() method test</title>
<meta name="assert" content="Inverts the playback rate of this player and seeks to the start of the source media if it has finished playing in the reversed direction">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#widl-AnimationPlayer-reverse-void">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
function testFinishedTrue(t, player) {
    setTimeout(function() {
        t.step(function () {
            assert_true(player.finished, 'AnimationPlayer should be finished');
            assert_equals(player.onFinishCounter, 1, 'AnimationPlayer.onfinish callback ' +
                 'shoul be called');
            t.done();
        });
    }, 40);
}

function testFinishedFalse(t, player) {
    setTimeout(function() {
        t.step(function () {
            assert_false(player.finished, 'AnimationPlayer should not be finished');
            assert_equals(player.onFinishCounter, 0, 'AnimationPlayer.onfinish callback ' +
                 'shoul not be called');
            t.done();
        });
    }, 40);
}

async_test(function(t) {
    var player = newPlayerFromAnimation(null);
    player.reverse();

    assert_equals(player.playbackRate, -1, 'Playback rate should be changed in sign');

    testFinishedTrue(t, player);
}, 'AnimationPlayer.reverse() reverses the player with no source content');

async_test(function(t) {
    var player = newPlayerFromAnimation(null);
    player.pause();
    player.reverse();

    assert_equals(player.playbackRate, -1, 'Playback rate should be changed in sign');
    assert_false(player.paused, 'AnimationPlayer should not be paused');

    testFinishedTrue(t, player);
}, 'AnimationPlayer.reverse() reverses paused player with no source content');

async_test(function(t) {
    var player = newPlayerFromAnimation(null);
    player.playbackRate = -0.4;
    player.pause();
    player.reverse();

    assert_equals(player.playbackRate, 0.4, 'Playback rate should be changed in sign');
    assert_false(player.paused, 'AnimationPlayer should not be paused');

    testFinishedTrue(t, player);
}, 'AnimationPlayer.reverse() reverses paused player with no source content and negative ' +
    'playback rate');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    player.reverse();

    assert_equals(player.playbackRate, -1, 'Playback rate should be changed in sign');

    testFinishedTrue(t, player);
}, 'AnimationPlayer.reverse() reverses player with source content');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    player.pause();
    player.reverse();

    assert_equals(player.playbackRate, -1, 'Playback rate should be changed in sign');
    assert_false(player.paused, 'AnimationPlayer should not be paused');

    testFinishedTrue(t, player);
}, 'AnimationPlayer.reverse() reverses paused player with source content');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    player.playbackRate = -0.4;
    player.pause();
    player.reverse();

    assert_equals(player.playbackRate, 0.4, 'Playback rate should be changed in sign');
    assert_false(player.paused, 'AnimationPlayer should not be paused');

    testFinishedFalse(t, player);
}, 'AnimationPlayer.reverse() reverses paused player with source content and negative ' +
    'playback rate');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    var timesToTest = [
        1,
        100,
        player.source.endTime / 2,
        player.source.endTime - 100,
        player.source.endTime - 1];

    timesToTest.forEach(function(time) {
        player.currentTime = time;
        player.pause();
        var playbackRate = player.playbackRate; 
        player.reverse();

        assert_equals(player.playbackRate, -1 * playbackRate, 'Playback rate should be ' +
            'changed in sign');
        assert_false(player.paused, 'AnimationPlayer should not be paused');
    });
    testFinishedFalse(t, player);
}, 'AnimationPlayer.reverse() reverses player when player is not limited');

test(function() {
    var player = newAnimationPlayer(createDiv(this));
    player.pause();
    player.playbackRate = 0;

    var timesToTest = [
        -100,
        -1,
        0,
        1,
        100,
        player.source.endTime / 2,
        player.source.endTime - 100,
        player.source.endTime - 1,
        player.source.endTime,
        player.source.endTime + 1,
        player.source.endTime + 100];

    timesToTest.forEach(function(time) {
        player.currentTime = time;
        player.reverse();

        assert_equals(player.currentTime, time, 'Seek should not be performed');
        assert_true(player.paused, 'AnimationPlayer paused state should not be changed');
    });
}, 'Test that if playback rate is zero reverse() call should do nothing');

test(function() {
    var player = newAnimationPlayer(createDiv(this));
    player.currentTime = player.source.endTime + 100;
    player.reverse();

    assert_equals(player.currentTime, player.source.endTime,
        'Seek to source content end time should be performed');
}, 'Test that if playback rate is > 0 and effective current time > source ' +
    'content\'s end time then seek to source end time');

test(function() {
    var player = newAnimationPlayer(createDiv(this));
    player.playbackRate = -1;
    player.currentTime = -100;
    player.reverse();

    assert_equals(player.currentTime, 0, 'Seek to zero time should be performed');
}, 'Test that if playback rate is < 0 and effective current time < 0 then seek to 0');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    setTimeout(function() {
        t.step(function() {
            player.reverse();
            setTimeout(function() {
                t.step(function() {
                    assert_approx_equals(player.currentTime, 0, EPSILON, 'Time should go ' +
                        'backward after reverse() call');
                    t.done();
                });
            }, 50);
        });
    }, 50);
}, 'Test that AnimationPlayer.reverse() call reverses player time');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    player.currentTime = player.source.endTime + 100;
    player.reverse();
    setTimeout(function() {
        t.step(function() {
            assert_approx_equals(player.currentTime, player.source.endTime - 50, EPSILON,
                'Time should go backward after reverse() call');
            t.done();
        });
    }, 50);
}, 'Test that AnimationPlayer.reverse() call reverses player time when current time > end time');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    player.currentTime = player.source.endTime / 2;
    player.playbackRate = -1;
    setTimeout(function() {
        t.step(function() {
            player.reverse();
            setTimeout(function() {
                t.step(function() {
                    assert_approx_equals(player.currentTime, player.source.endTime / 2, EPSILON,
                        'Time should go backward after reverse() call');
                    t.done();
                });
            }, 50);
        });
    }, 50);
}, 'Test that AnimationPlayer.reverse() call reverses player time when playback rate < 0');

async_test(function(t) {
    var player = newAnimationPlayer(createDiv(this));
    player.currentTime = -100;
    player.playbackRate = -1;
    player.reverse();
    setTimeout(function() {
        t.step(function() {
            assert_approx_equals(player.currentTime, 50, EPSILON,
                'Time should go backward after reverse() call');
            t.done();
        });
    }, 50);
}, 'Test that AnimationPlayer.reverse() call reverses player time when playback rate < 0 and ' +
    'current time < 0');
</script>
</body>
